requirements.txt
asgiref==3.11.0
Django==4.2.27
python-dotenv==1.2.1
sqlparse==0.5.5
